\name{df_ica}
\alias{df_ica}
\title{Gradient of projection index for independent component analysis.}
\description{
  Computes the gradient of the objective evaluating sample entropy of a data set projected onto a given vector.
  Only used within optimisation of projection. Not intended for alternative use.
}
\keyword{internal}
