\name{df_ppr}
\alias{df_ppr}
\title{Gradient of the projection index for projection pursuit regression}
\description{
  Computes the gradient of the loss of a smoother fit to projected covariates and corresponding response values.
  Only used within optimisation of projection. Not intended for alternative use.
}
\keyword{internal}
