\name{fk_dfmdh}
\alias{fk_dfmdh}
\title{Gradient of projection index for finding minimum density hyperplanes}
\description{
  Computes the gradient of the objective based on minimum integrated density of a hyperplane orthogonal to a given projection vector.
  Only used within optimisation of projection. Not intended for alternative use.
}
\keyword{internal}
