\name{plot.fk_density}
\alias{plot.fk_density}
\title{Plot method for class fk_density}
\description{
  Plot method for kernel density estimate obtained with function fk_density()
}
\usage{
\method{plot}{fk_density}(x, main = NULL, ...)
}
\arguments{
  \item{x}{an object of class fk_density, output from the function of the same name.}
  \item{main}{(optional) main title for plot.}
  \item{...}{(optional) further plotting parameters.}
}
\keyword{file}
