\name{plot.fk_ICA}
\alias{plot.fk_ICA}
\title{Plot method for class fk_ICA}
\description{
  Plot method for Independent Component Analysis model obtained with function fk_ICA()
}
\usage{
\method{plot}{fk_ICA}(x, ...)
}
\arguments{
  \item{x}{an object of class fk_ICA, output from the function of the same name.}
  \item{...}{(optional) further plotting parameters.}
}
\keyword{file}
