\name{plot.fk_mdh}
\alias{plot.fk_mdh}
\title{Plot method for class fk_mdh}
\description{
  Plot method for minimum density hyperplane solution obtained with function fk_mdh()
}
\usage{
\method{plot}{fk_mdh}(x, ...)
}
\arguments{
  \item{x}{an object of class fk_mdh, output from the function of the same name.}
  \item{...}{(optional) further plotting parameters.}
}
\keyword{file}
