\name{plot.fk_ppr}
\alias{plot.fk_ppr}
\title{Plot method for class fk_ppr}
\description{
  Plot method for projection pursuit regression model fit with function fk_ppr()
}
\usage{
\method{plot}{fk_ppr}(x, term = NULL, ...)
}
\arguments{
  \item{x}{an object of class fk_ppr, output from the function of the same name.}
  \item{term}{(optional) the term in the additive regression model to be visualised. If omitted then
  the residuals and fitted values for the full model are shown.}
  \item{...}{(optional) further arguments passed to or from other methods.}
}
\keyword{file}
