\name{plot.fk_regression}
\alias{plot.fk_regression}
\title{Plot method for class fk_regression}
\description{
  Plot method for kernel regression estimate obtained with function fk_regression()
}
\usage{
\method{plot}{fk_regression}(x, main = NULL, ...)
}
\arguments{
  \item{x}{an object of class fk_regression, output from the function of the same name.}
  \item{main}{(optional) main title for plot.}
  \item{...}{(optional) further plotting parameters.}
}
\keyword{file}
