\name{print.fk_density}
\alias{print.fk_density}
\title{Print method for class fk_density}
\description{
  Print method for kernel density estimate obtained with function fk_density()
}
\usage{
\method{print}{fk_density}(x, ...)
}
\arguments{
  \item{x}{an object of class fk_density, output from the function of the same name.}
  \item{...}{(optional) further arguments passed to or from other methods.}
}
\keyword{file}
