\name{print.fk_ICA}
\alias{print.fk_ICA}
\title{Print method for class fk_ICA}
\description{
  Print method for Independent Component Analysis model fit with function fk_ICA()
}
\usage{
\method{print}{fk_ICA}(x, ...)
}
\arguments{
  \item{x}{an object of class fk_ICA, output from the function of the same name.}
  \item{...}{(optional) further arguments passed to or from other methods.}
}
\keyword{file}
