\name{print.fk_mdh}
\alias{print.fk_mdh}
\title{Print method for class fk_mdh}
\description{
  Print method for minimum density hyperplane solution obtained with function fk_mdh()
}
\usage{
\method{print}{fk_mdh}(x, ...)
}
\arguments{
  \item{x}{an object of class fk_mdh, output from the function of the same name.}
  \item{...}{(optional) further arguments passed to or from other methods.}
}
\keyword{file}
