\name{print.fk_ppr}
\alias{print.fk_ppr}
\title{Print method for class fk_ppr}
\description{
  Print method for projection pursuit regression model fit with function fk_ppr()
}
\usage{
\method{print}{fk_ppr}(x, ...)
}
\arguments{
  \item{x}{an object of class fk_ppr, output from the function of the same name.}
  \item{...}{(optional) further arguments passed to or from other methods.}
}
\keyword{file}
