\name{print.fk_regression}
\alias{print.fk_regression}
\title{Print method for class fk_regression}
\description{
  Print method for kernel regression estimate obtained with function fk_regression()
}
\usage{
\method{print}{fk_regression}(x, ...)
}
\arguments{
  \item{x}{an object of class fk_regression, output from the function of the same name.}
  \item{...}{(optional) further arguments passed to or from other methods.}
}
\keyword{file}
