\name{sm_bin_wts}
\alias{sm_bin_wts}
\title{Compute smoothed bin weights}
\description{
  Computes total weight from an initial weight vector falling into regular bins, based on corresponding location of sample points.
  Individual points are shared between pairs of adjacent bins to produce slightly smoother estimation. Not intended for general use.
}
\keyword{internal}
