\name{whiten}
\alias{whiten}
\title{Whitening (standardising) a data matrix}
\description{
  Transforms data matrix by centering, projecting onto its first k principal components, and standardising the resulting components. The output is a zero mean, identity covariance data matrix. Only used for pre-processing prior to ICA. Not intended for alternative access.
}
\keyword{internal}
